+++ demos/layout/designMenuCss.tsx
@@ -192,7 +192,6 @@ const Demo = () => {
       </ProLayout>
       <SettingDrawer
         pathname={pathname}
-        enableDarkTheme
         getContainer={() => document.getElementById('test-pro-layout')}
         settings={settings}
         onSettingChange={(changeSetting) => {