+++ demos/layout/designSiderMenu.tsx
@@ -92,7 +92,6 @@ const Demo = () => {
       </ProLayout>
       <SettingDrawer
         pathname={pathname}
-        enableDarkTheme
         getContainer={() => document.getElementById('test-pro-layout')}
         settings={settings}
         onSettingChange={(changeSetting) => {